+++ tests/unit_tests/test_line.py
@@ -92,7 +92,7 @@ def test_calculate_depth() -> None:
     assert (close_paren_n.depth, close_paren_n.open_brackets) == ((1, 0), [select_n])
 
 
-def test_bare_line(source_string: str, bare_line: Line) -> None:
+def test_bare_line(bare_line: Line) -> None:
     assert str(bare_line) == ""
 
     assert not bare_line.starts_with_unterm_keyword
@@ -105,6 +105,8 @@ def test_bare_line(source_string: str, bare_line: Line) -> None:
     assert not bare_line.is_standalone_multiline_node
     assert not bare_line.is_too_long(88)
     assert not bare_line.opens_new_bracket
+    assert bare_line.open_brackets == []
+    assert bare_line.open_jinja_blocks == []
 
 
 def test_simple_line(
@@ -161,6 +163,12 @@ def test_bare_append_newline(bare_line: Line) -> None:
     new_last_node = bare_line.nodes[-1]
     assert new_last_node.token.type == TokenType.NEWLINE
     assert (new_last_node.token.spos, new_last_node.token.epos) == (0, 0)
+
+
+def test_bare_with_previous_open_lists(bare_line: Line, simple_line: Line) -> None:
+    bare_line.previous_node = simple_line.nodes[-1]
+    assert bare_line.open_jinja_blocks == simple_line.nodes[-1].open_jinja_blocks
+    assert bare_line.open_brackets == simple_line.nodes[-1].open_brackets
 
 
 def test_bare_with_previous_append_newline(bare_line: Line, simple_line: Line) -> None:
@@ -326,6 +334,32 @@ def test_is_standalone_multiline_node(bare_line: Line, simple_line: Line) -> Non
     assert not simple_line.is_standalone_multiline_node
 
 
+def test_is_standalone_jinja_statement(bare_line: Line, simple_line: Line) -> None:
+
+    assert not bare_line.is_standalone_jinja_statement
+    assert not simple_line.is_standalone_jinja_statement
+
+    jinja_statement = Token(
+        type=TokenType.JINJA_STATEMENT,
+        prefix="",
+        token="{% my JINJA %}",
+        spos=0,
+        epos=14,
+    )
+
+    bare_line.append_token(jinja_statement)
+    simple_line.append_token(jinja_statement)
+
+    assert bare_line.is_standalone_jinja_statement
+    assert not simple_line.is_standalone_jinja_statement
+
+    bare_line.append_newline()
+    simple_line.append_newline()
+
+    assert bare_line.is_standalone_jinja_statement
+    assert not simple_line.is_standalone_jinja_statement
+
+
 def test_calculate_depth_exception() -> None:
 
     close_paren = Token(
@@ -388,6 +422,15 @@ def test_identifier_whitespace(default_mode: Mode, source_string: str) -> None:
     ).parse_query(source_string=source_string)
     parsed_string = "".join(str(line) for line in q.lines)
     assert source_string == parsed_string
+
+
+def test_jinja_whitespace_after_newline(default_mode: Mode) -> None:
+    source_string = "select *\nfrom {{model}}\nwhere {{ column_name }} < 0\n"
+    q = default_mode.dialect.initialize_analyzer(
+        line_length=default_mode.line_length
+    ).parse_query(source_string=source_string)
+    where_node = [node for node in q.nodes if node.value == "where"][0]
+    assert where_node.prefix == " "  # one space
 
 
 def test_capitalization(default_mode: Mode) -> None:
@@ -513,10 +556,10 @@ def test_jinja_depth(default_mode: Mode) -> None:
         (1, 2),  # \n
         (0, 2),  # union all
         (1, 2),  # \n
-        (1, 1),  # {%- endif %}
-        (1, 1),  # \n
-        (1, 0),  # {% endfor %}
-        (1, 0),  # \n
+        (0, 1),  # {%- endif %}
+        (0, 1),  # \n
+        (0, 0),  # {% endfor %}
+        (0, 0),  # \n
     ]
     actual = [node.depth for node in q.nodes]
     assert actual == expected